Drop table defs 2.0 part 1
- Start building drop table defs from ts files that make use of deepsight enums
- Add traversal sections to "encounters" array
- Use directives provided by Charm for encounter names
- Add guesses for all phase hashes cross referenced between milestones (for raids), insertion points (all activities), and directives provided by Charm
- Fix weekly dungeon not showing if first character has already completed it
- Mark some phases as traversal phases so they're not considered normal encounters
- Improve enum generation a lot by excluding a ton of stuff from it
- Fix a bunch of cases where enum differences could then be simplified down into being identical again
-

+++ src/ui/inventory/tooltip/ItemTooltipSource.ts
@@ -73,7 +73,6 @@ export default class ItemTooltipSource extends Component {
 				.classes.add(ItemTooltipSourceClasses.Activity)
 				.style.set("--icon", undefined
 					?? Display.icon(source.dropTable.displayProperties)
-					?? Display.icon(source.record)
 					?? Display.icon(activity))
 				.appendTo(this.activityWrapper);
 
@@ -172,8 +171,13 @@ export default class ItemTooltipSource extends Component {
 		if (!source.dropTable.encounters?.length)
 			return;
 
+		let realEncounterIndex = 0;
 		for (let i = 0; i < source.dropTable.encounters.length; i++) {
 			const encounter = source.dropTable.encounters[i];
+			if (encounter.traversal)
+				continue;
+
+			realEncounterIndex++;
 			const dropTable = encounter.dropTableMergeStrategy === "replace" ? encounter.dropTable
 				: { ...source.dropTable.dropTable, ...encounter.dropTable };
 			if (!dropTable?.[item.definition.hash])
@@ -185,7 +189,7 @@ export default class ItemTooltipSource extends Component {
 
 			Component.create()
 				.classes.add(ItemTooltipSourceClasses.ActivityPhaseIndex)
-				.text.set(`${i + 1}`)
+				.text.set(`${realEncounterIndex}`)
 				.appendTo(phaseComponent);
 
 			Component.create()
